Changed bl_category from "Proxy & Cache" to just "Proxy"
Newer versions of blender split the old "Proxy & Cache" tab into separate "Proxy" and "Cache" tabs. Installing this add-on as-is in Blender 3.0 creates a 3rd tab named "Proxy & Cache".

For simplicity sake, I've moved this add-on into the "Proxy" category/tab so that it behaves more consistently like it used to.